requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.7.0
certifi==2023.7.22
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
contourpy==1.2.0
cookiecutter==1.7.3
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.2.0
dj-rest-auth==7.0.1
Django==4.2.23
django-active-link==0.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
django-phonenumber-field==8.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
env==0.1.0
et_xmlfile==2.0.0
executing==1.2.0
fast2sms==1.0.0
feedparser==6.0.10
filelock==3.13.1
Flask==3.0.0
fonttools==4.45.0
fpdf==1.7.2
frozenlist==1.4.0
funcs==0.10.0
googletrans==3.0.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpie==3.2.2
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jedi==0.19.0
Jinja2==3.1.2
jinja2-time==0.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
MouseInfo==0.1.3
multidict==6.0.4
multipledispatch==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.4
named==1.4.2
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==23.1
pafy==0.5.5
pandas==2.1.2
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
poyo==0.5.0
prettytable==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.16.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pyjokes==0.6.0
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyotp==2.9.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-slugify==8.0.4
python3-openid==3.2.0
pyttsx3==2.90
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
pywin32-ctypes==0.2.2
PyYAML==3.13
pyzmq==25.1.1
razorpay==1.4.2
renv==0.3.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
seaborn==0.13.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
solus==1.2.2
soupsieve==2.5
SpeechRecognition==3.10.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
tornado==6.3.3
traitlets==5.10.1
twilio==8.10.1
types-python-dateutil==2.8.19.20240106
typing==3.7.4.3
typing-aliases==1.10.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.7
vic==0.0.1
virtualenv==20.25.0
wcwidth==0.2.6
Werkzeug==3.0.1
whitenoise==5.2.0
wikipedia==1.4.0
winshell==0.6
wolframalpha==5.0.0
wraps==0.9.2
xlwt==1.3.0
xmltodict==0.13.0
yarl==1.9.2
